use general proposal;
